% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_pdf_download.R
\name{article_pdf_download}
\alias{article_pdf_download}
\title{Batch download}
\usage{
article_pdf_download(infilepath, outfilepath, match = NULL)
}
\arguments{
\item{infilepath}{A character; path to target folder with input files}

\item{outfilepath}{A character; path to folder for export files}

\item{match}{A file that provides titles to match; designed to be output of Colandr}
}
\description{
This function processes and batch downloads files
}
\examples{
article_pdf_download()
}
\keyword{download}
