% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsevier_pdf_download.R
\name{elsevier_pdf_download}
\alias{elsevier_pdf_download}
\title{Elsevier check}
\usage{
elsevier_pdf_download(elsevier_xml_link, filepath)
}
\arguments{
\item{elsevier_xml_link}{A character; elsevier link of type xml as returned by crminer's crmlinks() function}

\item{filepath}{A character; path to desired output location for downloaded PDF}
}
\value{
Nothing. Download PDFs to HD
}
\description{
This function uses webscraping techniques to download a PDF file from elsevier's server
}
\examples{
\dontrun{
elsevier_pdf_download('https://api.elsevier.com/content/article/PII:0167198795004585?httpAccept=text/xml','/Users/timothy/Documents/soilc-text_mapping/data')
}
}
\keyword{elsevier}
