% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsevier_tagger.R
\name{elsevier_tagger}
\alias{elsevier_tagger}
\title{Tag papers from the Elsevier publisher}
\usage{
elsevier_tagger(df_doi, col_links)
}
\arguments{
\item{df_doi}{(data frame) DOI links}

\item{col_links}{(character) column name countaining the links}
}
\value{
dataframe
}
\description{
Tag papers from the Elsevier publisher
}
\examples{
elsevier_tagger(my_df, "links")
}
