% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_to_doi.R
\name{title_to_doi}
\alias{title_to_doi}
\title{Title-DOI Matcher}
\usage{
title_to_doi(papers, bib.dir)
}
\arguments{
\item{papers}{A .csv exported from Colandr}

\item{bib.dir}{A directory containing all .bib files imported to Colandr}
}
\value{
A data frame of titles, journals, authors, and DOIs in the .bib format
}
\description{
This function matches titles exported from Colandr to .bib files
and exports the associated DOI from Web of Science
}
\examples{
title_to_doi()
}
\keyword{Colandr}
