% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_pdf_download.R
\name{article_pdf_download}
\alias{article_pdf_download}
\title{Batch download articles from bibtex files}
\usage{
article_pdf_download(infilepath, outfilepath = infilepath, colandr = NULL,
  cond = "included")
}
\arguments{
\item{infilepath}{A character; path to target folder with input files}

\item{outfilepath}{A character; path to folder for export files}

\item{colandr}{A file that provides titles to match; designed to be output of Colandr}

\item{cond}{Condition that defines sorting of output from Colandr file}
}
\value{
data frame containing dowload information
}
\description{
Batch download articles from bibtex files
}
\examples{
article_pdf_download(infilepath = "/data/isi_searches/", outfilepath = "data")
}
