% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_reader.R
\name{bibfile_reader}
\alias{bibfile_reader}
\title{Read DOIs/Metadata from bib files}
\usage{
bibfile_reader(isi_dir)
}
\arguments{
\item{isi_dir}{(character) Path to target folder with bibtex files}
}
\value{
data frame witht eh bibtex file content
}
\description{
Read DOIs/Metadata from bib files
}
\examples{
\dontrun{bibfile_reader("data/isi_search")}
}
