% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_binary.R
\name{is_binary}
\alias{is_binary}
\title{Binary check function}
\usage{
is_binary(filepath, max = 1000)
}
\arguments{
\item{filepath}{A character; path to target file}

\item{max}{An integer; max number of characters in file to be checked (default 1000)}
}
\value{
A boolean; TRUE if file is binary and FALSE otherwise
}
\description{
Adatapted from: https://stackoverflow.com/questions/16350164/native-method-in-r-to-test-if-file-is-ascii
}
\details{
This function checks if a file is binary
}
\examples{
\dontrun{is_binary()}
}
