% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_to_doi.R
\name{title_to_doi}
\alias{title_to_doi}
\title{Title-DOI Matcher}
\usage{
title_to_doi(papers, file_list, condition)
}
\arguments{
\item{papers}{A csv file exported from Colandr}

\item{file_list}{A tibble of .bib files imported to Colandr}

\item{condition}{Condition for sorting papers}
}
\value{
A data frame of titles, journals, authors, and DOIs in the .bib format
}
\description{
This function matches titles exported from Colandr to .bib files
and exports the associated DOI from Web of Science
}
\examples{
\dontrun{
title_to_doi(papers, df_citations, cond)
}

}
\keyword{Colandr}
